+++ src/main/java/duke/Commands/InvalidCommand.java
@@ -1,10 +1,10 @@
-package Commands;
+package duke.Commands;
 
-import Exceptions.DukeException;
-import Exceptions.InvalidInputException;
-import Storage.Storage;
-import TaskList.TaskList;
-import Ui.Ui;
+import duke.Exceptions.DukeException;
+import duke.Exceptions.InvalidInputException;
+import duke.Storage.Storage;
+import duke.TaskList.TaskList;
+import duke.Ui.Ui;
 
 public class InvalidCommand implements Command {
     @Override
