+++ src/main/java/duke/Commands/TaskCommand.java
@@ -1,18 +1,18 @@
-package Commands;
+package duke.Commands;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-import Exceptions.DukeException;
-import Exceptions.InvalidInputException;
-import Exceptions.ListException;
-import Storage.Storage;
-import Task.Deadline;
-import Task.Event;
-import Task.Todo;
-import TaskList.TaskList;
-import Ui.Ui;
+import duke.Exceptions.DukeException;
+import duke.Exceptions.InvalidInputException;
+import duke.Exceptions.ListException;
+import duke.Storage.Storage;
+import duke.Task.Deadline;
+import duke.Task.Event;
+import duke.Task.Todo;
+import duke.TaskList.TaskList;
+import duke.Ui.Ui;
 
 public class TaskCommand implements Command {
