+++ src/main/java/duke/Exceptions/DukeException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package duke.Exceptions;
 
 public class DukeException extends Exception {
     public DukeException(String message) {
