+++ src/main/java/duke/Exceptions/InvalidInputException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package duke.Exceptions;
 
 public class InvalidInputException extends DukeException {
     public InvalidInputException(String message) {
