+++ src/main/java/duke/Task/Deadline.java
@@ -1,4 +1,4 @@
-package Task;
+package duke.Task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
