+++ src/main/java/duke/Task/Event.java
@@ -1,4 +1,4 @@
-package Task;
+package duke.Task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
